Extracts experience ui config from experience data
