+++ src/data/experience.js
@@ -7,7 +7,6 @@ import {
   NET_DEVELOPER_COMKNOW,
   IOS_APPRENTICE_REEA,
 } from '../constants/experience'
-
 import {
   ProgrammingLanguage,
   FrontEndTech,
@@ -17,19 +16,6 @@ import {
   Cloud,
   OtherTools,
 } from './technologies'
-
-/*
-This configuration data is used to display the experience section on the front-end.
-In a way that similar experiences such as .Net Developer at Branded3 and ComKnow are merged.
-*/
-const experience_ui_config = [
-  [MENTOR_SCHOOL_OF_CODE.id],
-  [SOFTWARE_ENGINEER_BJSS.id],
-  [FULL_STACK_DEVELOPER_RIGHTINDEM.id],
-  [CO_FOUNDER_WRAPTIME.id],
-  [NET_DEVELOPER_BRANDED3.id, NET_DEVELOPER_COMKNOW.id],
-  [IOS_APPRENTICE_REEA.id],
-]
 
 export const experience = [
   {
